requirements.txt
azure-identity==1.15.0
# Flask[async]==2.3.2
azure-storage-blob==12.17.0
python-dotenv==1.0.0
azure-cosmos==4.5.0
quart==0.19.4
uvicorn==0.24.0
aiohttp==3.9.2
gunicorn==20.1.0
pydantic-settings==2.2.1
openai==1.44.1
azure-search-documents==11.5.1
instructor==1.4.1
